CM-41 - move field transform methods to separate FieldTransformer class
