+++ src/main/java/uk/gov/companieshouse/uri/web/model/CompanyDetails.java
@@ -6,7 +6,9 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public class CompanyDetails {
+import uk.gov.companieshouse.uri.web.transformer.FieldTransformer;
+
+public class CompanyDetails extends FieldTransformer {
 
     @JsonProperty(value = "CompanyName")
     private String companyName;
@@ -163,35 +165,6 @@ public class CompanyDetails {
 
     public void setMortgageTotals(MortgageTotals mortgageTotals) {
         this.mortgageTotals = mortgageTotals;
-    }
-    
-    /**
-     * Converts a string date of format 'DD/MM/YYYY' to 'YYYY-MM-DD'
-     * for use in XML template output
-     * If the supplied date is null or is not the correct format, then
-     * the supplied string is returned unaltered
-     * @param displayDateString - date string of the format 'DD/MM/YYYY'
-     */    
-    public String xmlDate(String displayDateString) {
-        if (displayDateString != null && displayDateString.matches("^[0-9]{2}/[0-9]{2}/[0-9]{4}$")) {
-            StringBuilder builder = new StringBuilder();
-            builder.append(displayDateString.substring(6));
-            builder.append('-');
-            builder.append(displayDateString.substring(3, 5));
-            builder.append('-');
-            builder.append(displayDateString.substring(0, 2));
-            return builder.toString();
-        }
-        return displayDateString;
-    }
-    
-    /**
-     * Escapes a string so that it is suitable for csv output
-     * This is used to handle fields that may contain one or more " characters
-     * @param unescapedString - string that may have " present
-     */    
-    public String csvEscape(String unescapedString) {
-        return unescapedString == null ? null : unescapedString.replaceAll("\"", "\"\"");
     }
 
     @Override
